Add aws_profile config option to charon.yaml
This is the first part. The second part will address needed changes in storage.py

Feat: Add AWS profile retrieval from envar / config

Extract AWS profile at this level so we can get it from config if it's not set as an environment variable. Then, pass it to the various `handle_*` methods.

Accept AWS profile param for handle_* and pass to S3Client

Accept aws_profile param in handle_* and pass to S3Client

Pass in AWS profile instead of reading from envar

Allow calling code to pass in the aws_profile, so we have a chance to read it from charon.yaml or the environment variable, instead of being locked into reading from envar.